[ADD] extract PostController for better state management and scroll handling
